Refactoring.

+++ YasEngineSoftware/CosinusPointsGenerator.hpp
@@ -6,7 +6,7 @@
 class CosinusPointsGenerator : public PointsGenerator
 {
 	public:
-        static std::vector<float> numbers;
+        static std::vector<double> numbers; // FOR GAMEPLAY PURPOSE!!
 		virtual PointsSet* generatePoints() override;
 
 };
